Create 2ndHighestSalary.py
Learning that working with python dataframes and python series can be cumbersome and annoying. It's mostly slogging through documentation in the initial learning curve and familiarizing myself with slicing datasets : row-wise dataframe slicing or col-wise dataframe slicing. Additionally learning that I can use index operator access patterns for series ( and I'm not restricted to iLoc ), but I have to be careful with the bool parameters passed in methods ( make sure it's correct ).