Create LastTransaction.py
Fun problem with a group by and a max operation
Leverage the `.transform(aggr)` capability to add a column in place and minimize amount of data frames generated in memory or returned in code. In-place data frame modification more ideal
Additionally, leverage a filter on a max for group by to get the final result
Also exert caution on the need for pd.datetime(...) conversions. pd.datetime(...) on timestamp values seems to frequently arise