Create top3Users.py
Woah : the `rank()` window function is super fast to code up with Pandas Dataframes. Also the ability to leverage the .groupBy(...) followed up by .head(....) to select the top N in each group is super useful. Just take note of a need to sort before the group by and head operations too!