Update bestPerformingAdvertisers.py
The problem in itself was not actually as "difficult" as it seemed ... it was the understanding of usage of `isoCalendar` dates versus `non-isoCalendar` dates in the aggregations, which led to a one-off error in a calculation of an aggregation of the max weekly revenue amounts of advertiser transactions. 
Wow. TYTY my Google L5 TL back in the day for knowing to catch the isodate issues ahead of time!